added responsiveness
