+++ src/Components/Contact Us/Dropdown.jsx
@@ -9,7 +9,7 @@ const Dropdown = ({ selectedCounty, handleCountyChange }) => {
   return (
     <div>
       <select
-        className="text-black w-[15rem] h-12 "
+        className="text-black h-12 my-4 w-full md:w-[15rem]"
         id="drop"
         onChange={handleDropdownChange}
         value={selectedCounty}
